// src/app/core/models/person.ts
export class Person {
  email: string = "";
  name: string = "";
  surname: string = "";
  roles: string[] = [];
  display_name: string = "";
  description: string = "";

  id: string = "";
  version: number = 0;
  tenant_id: string = "";
  location: string = "";
  resource_type: string = "";
  created_at: string = "";
  updated_at: string = "";
  created_by: string = "";
  updated_by: string = "";
}